+++ services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/status/suspend/SuspendUserCommand.java
@@ -1,10 +1,10 @@
-package io.ciphernance.identity.application.command.user.block;
+package io.ciphernance.identity.application.command.user.status.suspend;
 
 import io.ciphernance.identity.application.mediator.Command;
 
 import java.util.UUID;
 
-public record BlockUserCommand(
+public record SuspendUserCommand(
 
         UUID userId
 ) implements Command<Void> { }
